requirements.txt
pandas
pandasai
streamlit
python-dotenv